- Change argument to add constant matrix for assembly - update tests
